llamaindex/requirements.txt
llama-index==0.10.18
llama-index-llms-groq==0.1.3
chainlit==1.0.401
groq==0.4.2
python-dotenv==1.0.1
llama-index-embeddings-gemini==0.1.4
llama-index-postprocessor-cohere-rerank==0.1.2